Check for numbers in sentences can be switched off
If the CorporaCreator is used with data in which it is valid that
sentences contain numbers, there should be a way to allow them.

With the optional command line parameter "-c" this check can now be
skipped.

Usage:

create-corpora [other args] -c {true, false, t, f 0, 1, y, n, yes, no}